misc: rename centeredPage file
