+++ src/pages/settings/Dunnings/CreateDunning.tsx
@@ -6,7 +6,7 @@ import { array, boolean, number, object, string } from 'yup'
 
 import { Alert, Button, Tooltip, Typography } from '~/components/designSystem'
 import { AmountInputField, ComboBoxField, SwitchField, TextInputField } from '~/components/form'
-import { CenteredPage } from '~/components/layouts/Pages'
+import { CenteredPage } from '~/components/layouts/CenteredPage'
 import {
   DefaultCampaignDialog,
   DefaultCampaignDialogRef,
